Move realtime composition event handlers into their own file
