+++ internal/controller/apiextensions/composite/reconciler.go
@@ -26,12 +26,8 @@ import (
 
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	kunstructured "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/util/workqueue"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	runtimeevent "sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
@@ -730,48 +726,4 @@ func getComposerResourcesNames(cds []ComposedResource) []string {
 		names[i] = string(cd.ResourceName)
 	}
 	return names
-}
-
-// EnqueueForCompositionRevisionFunc returns a function that enqueues (the
-// related) XRs when a new CompositionRevision is created. This speeds up
-// reconciliation of XRs on changes to the Composition by not having to wait for
-// the 60s sync period, but be instant.
-func EnqueueForCompositionRevisionFunc(of resource.CompositeKind, list func(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error, log logging.Logger) func(ctx context.Context, createEvent runtimeevent.TypedCreateEvent[*v1.CompositionRevision], q workqueue.RateLimitingInterface) {
-	return func(ctx context.Context, createEvent runtimeevent.TypedCreateEvent[*v1.CompositionRevision], q workqueue.RateLimitingInterface) {
-		rev := createEvent.Object
-
-		// get all XRs
-		xrs := kunstructured.UnstructuredList{}
-		xrs.SetGroupVersionKind(schema.GroupVersionKind(of))
-		xrs.SetKind(schema.GroupVersionKind(of).Kind + "List")
-		if err := list(ctx, &xrs); err != nil {
-			// logging is most we can do here. This is a programming error if it happens.
-			log.Info("cannot list in CompositionRevision handler", "type", schema.GroupVersionKind(of).String(), "error", err)
-			return
-		}
-
-		// enqueue all those that reference the Composition of this revision
-		compName := rev.Labels[v1.LabelCompositionName]
-		if compName == "" {
-			return
-		}
-		for _, u := range xrs.Items {
-			xr := composite.Unstructured{Unstructured: u}
-
-			// only automatic
-			if pol := xr.GetCompositionUpdatePolicy(); pol != nil && *pol == xpv1.UpdateManual {
-				continue
-			}
-
-			// only those that reference the right Composition
-			if ref := xr.GetCompositionReference(); ref == nil || ref.Name != compName {
-				continue
-			}
-
-			q.Add(reconcile.Request{NamespacedName: types.NamespacedName{
-				Name:      xr.GetName(),
-				Namespace: xr.GetNamespace(),
-			}})
-		}
-	}
 }
