+++ internal/controller/apiextensions/definition/indexes.go
@@ -78,26 +78,3 @@ func IndexCompositeResourcesRefs(o client.Object) []string {
 func refKey(ns, name, kind, apiVersion string) string {
 	return fmt.Sprintf("%s.%s.%s.%s", name, ns, kind, apiVersion)
 }
-
-// TODO(negz): Figure out a way to plumb this with controller-runtime v0.18.x
-// style sources.
-
-// func enqueueXRsForMR(ca cache.Cache, xrGVK schema.GroupVersionKind, log logging.Logger) func(ctx context.Context, ev runtimeevent.UpdateEvent, q workqueue.RateLimitingInterface) {
-// 	return func(ctx context.Context, ev runtimeevent.UpdateEvent, q workqueue.RateLimitingInterface) {
-// 		mrGVK := ev.ObjectNew.GetObjectKind().GroupVersionKind()
-// 		key := refKey(ev.ObjectNew.GetNamespace(), ev.ObjectNew.GetName(), mrGVK.Kind, mrGVK.GroupVersion().String())
-//
-// 		composites := kunstructured.UnstructuredList{}
-// 		composites.SetGroupVersionKind(xrGVK.GroupVersion().WithKind(xrGVK.Kind + "List"))
-// 		if err := ca.List(ctx, &composites, client.MatchingFields{compositeResourcesRefsIndex: key}); err != nil {
-// 			log.Debug("cannot list composite resources related to a MR change", "error", err, "gvk", xrGVK.String(), "fieldSelector", compositeResourcesRefsIndex+"="+key)
-// 			return
-// 		}
-//
-// 		// queue those composites for reconciliation
-// 		for _, xr := range composites.Items {
-// 			log.Info("Enqueueing composite resource because managed resource changed", "name", xr.GetName(), "mrGVK", mrGVK.String(), "mrName", ev.ObjectNew.GetName())
-// 			q.Add(reconcile.Request{NamespacedName: types.NamespacedName{Name: xr.GetName()}})
-// 		}
-// 	}
-// }
